move netdataapi pkg to agent
